Refactor PickL
