+++ PickL/ViewAdaptor/ViewAdaptor.swift
@@ -11,7 +11,7 @@ public final class ViewAdaptor: BaseAdaptor, SpecificAdaptor {
 
     // MARK: - UIPickerViewDelegate
 
-    public func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
+    func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
         return delegate.pickerView(pickerView, viewForRow: row, forComponent: component, reusing: view)
     }
 }
